Renaming package quickpay to quickpay_api_client
