+++ setup.py
@@ -10,7 +10,7 @@ reqs = ['requests>=2.5']
 tests_requires = ['nose', 'httpretty', 'mock']
 
 version = ''
-with open('quickpay/__init__.py', 'r') as fd:
+with open('quickpay_api_client/__init__.py', 'r') as fd:
     version = re.search(r'^__version__\s*=\s*[\'"]([^\'"]*)[\'"]', fd.read(),
                         re.MULTILINE).group(1)
 
@@ -18,13 +18,13 @@ if not version:
     raise RuntimeError('Cannot find version information')
 
 setup(
-    name='quickpay',
+    name='quickpay-api-client',
     version=version,
     description='Python client for QuickPay API',
     author_email="support@quickpay.net",
     author="QuickPay Developers",
     url="https://github.com/QuickPay/quickpay-python-client",
-    packages=['quickpay'],
+    packages=['quickpay_api_client'],
     install_requires=reqs,
     tests_requires=tests_requires,
     test_suite='nose.collector')
